// src/modules/emergency-room/assessment/models/assessment-ugd-models.ts
export interface IDaftarTebus {
    Catatan: string;
    ID_AturanPakai: string;
    ID_Satuan: string;
    Jumlah: string;
    Kode_AturanPakai: string;
    Kode_Obat: string;
    Nama_AturanPakai: string;
    Nama_Obat: string;
    Nama_Satuan: string;
  }

export interface IFormFarmasi {
    Daftar_Tebus: Array<IDaftarTebus>;
    Keterangan: string;
    Status_Tebus: string;
    Waktu_Tebus: string;
  }

export class FormFarmasi {
    Daftar_Tebus: Array<IDaftarTebus>;
    Keterangan: string;
    Status_Tebus: string;
    Waktu_Tebus: string;

    constructor(req: IFormFarmasi) {
      this.Daftar_Tebus = req.Daftar_Tebus;
      this.Keterangan = req.Keterangan;
      this.Status_Tebus = req.Status_Tebus;
      this.Waktu_Tebus = req.Waktu_Tebus;
    }
}

export interface IAssessmentUgdModel extends IDataModel  {
    form:  IAssessmentUgdFormModel;
    formFarmasi: IFormFarmasi;
    isDefault: boolean;
    obat: Array<IMedicine>;
    aturan_pakai: Array<IHowToUse>;
    dokter: Array<IDoctorModel>;
}

export class AssessmentUgdModel extends DataModel {
    form:  IAssessmentUgdFormModel;
    formFarmasi: IFormFarmasi;
    isDefault: boolean;
    obat: Array<IMedicine>;
    aturan_pakai: Array<IHowToUse>;
    dokter: Array<IDoctorModel>;

    constructor(assessment: IAssessmentUgdModel) {
      super(assessment);
      this.form = assessment.form;
      this.formFarmasi = assessment.formFarmasi;
      this.isDefault = assessment.isDefault;
      this.obat = (Array.isArray(assessment.obat)) ? assessment.obat.map((a) => new Medicine(a)) : [];
      this.aturan_pakai = (Array.isArray(assessment.aturan_pakai)) ? assessment.aturan_pakai.map((a) => new HowToUse(a)) : [];
      this.dokter = (Array.isArray(assessment.dokter)) ? assessment.dokter.map((a) => new DoctorModel(a)) : [];
    }
}